+++ src/init.c
@@ -11,7 +11,7 @@ extern SEXP R_make_big_file(SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
   { "R_zip_list",      (DL_FUNC) &R_zip_list,      1 },
-  { "R_zip_zip",       (DL_FUNC) &R_zip_zip,       6 },
+  { "R_zip_zip",       (DL_FUNC) &R_zip_zip,       7 },
   { "R_make_big_file", (DL_FUNC) &R_make_big_file, 2 },
   { NULL, NULL, 0 }
 };