+++ NEWS.md
@@ -4,6 +4,9 @@
 * New functions: `zipr()` and `zipr_append()`, that always store
   relative file names in the archive
 
+* Add time stamps to zip file entries, and include them in the
+  output of `zip_list()`
+
 * Include directories in the archives, empty ones as well
 
 * Fix reporting the size of large files in the archive